refactor settings service
